Terminal styling
I wanted to convert this immediately to a UI program, but I have a lot to learn about event driven programming before I can do this to my satisfaction. Therefore, I decided to finish this version off as a terminal only program. This version uses pyfiglet and clint.textui to style some text in the terminal. I will continue adding functionality in the future.